Karma-Polymer now works in a Polymer 2.0 project :-)
I believe this should work as well in a Polymer 1.0 or vanilla Web
Components project as well

In order to make this work, we need careful control over the load order
of the polyfill, source files, dependencies and tests, so I added new
config options to provide that control and updated the documentation
appropriately.

Note that polymer.html does NOT need to be included anymore (as any
components that use it should pull it in themselves via HTML imports),
so I removed the documentation around that.

Replaced "jasmine" with "mocha" in documentation, as that is what
web-component-tester uses, so most developers will probably already have
tests in mocha.

This obviously needs a major version bump, so updating documentation as
well to reflect the assumption the version number will be 1.0.0 when
these changes are released